+++ packages/transaction-controller/src/TransactionController.ts
@@ -139,7 +139,12 @@ import {
 } from './utils/eip7702';
 import { validateConfirmedExternalTransaction } from './utils/external-transactions';
 import { updateFirstTimeInteraction } from './utils/first-time-interaction';
-import { addGasBuffer, estimateGas, updateGas } from './utils/gas';
+import {
+  addGasBuffer,
+  estimateGas,
+  estimateGasBatch,
+  updateGas,
+} from './utils/gas';
 import {
   checkGasFeeTokenBeforePublish,
   getGasFeeTokens,
@@ -310,6 +315,11 @@ export type TransactionControllerEstimateGasAction = {
   handler: TransactionController['estimateGas'];
 };
 
+export type TransactionControllerEstimateGasBatchAction = {
+  type: `${typeof controllerName}:estimateGasBatch`;
+  handler: TransactionController['estimateGasBatch'];
+};
+
 /**
  * Adds external provided transaction to state as confirmed transaction.
  *
@@ -400,6 +410,7 @@ export type TransactionControllerActions =
   | TransactionControllerAddTransactionBatchAction
   | TransactionControllerConfirmExternalTransactionAction
   | TransactionControllerEstimateGasAction
+  | TransactionControllerEstimateGasBatchAction
   | TransactionControllerGetGasFeeTokensAction
   | TransactionControllerGetNonceLockAction
   | TransactionControllerGetStateAction
@@ -1795,6 +1806,39 @@ export class TransactionController extends BaseController<
     });
 
     return { gas: estimatedGas, simulationFails };
+  }
+
+  /**
+   * Estimates required gas for a batch of transactions.
+   *
+   * @param request - Request object.
+   * @param request.chainId - Chain ID of the transactions.
+   * @param request.from - Address of the sender.
+   * @param request.transactions - Array of transactions within a batch request.
+   * @returns Object containing the gas limit.
+   */
+  async estimateGasBatch({
+    chainId,
+    from,
+    transactions,
+  }: {
+    chainId: Hex;
+    from: Hex;
+    transactions: BatchTransactionParams[];
+  }): Promise<{ totalGasLimit: number; gasLimits: number[] }> {
+    const ethQuery = this.#getEthQuery({
+      chainId,
+    });
+
+    return estimateGasBatch({
+      chainId,
+      ethQuery,
+      from,
+      getSimulationConfig: this.#getSimulationConfig,
+      isAtomicBatchSupported: this.isAtomicBatchSupported.bind(this),
+      messenger: this.messenger,
+      transactions,
+    });
   }
 
   /**
@@ -3440,15 +3484,29 @@ export class TransactionController extends BaseController<
     transactionMeta: TransactionMeta,
     { skipSubmitHistory }: { skipSubmitHistory?: boolean } = {},
   ): Promise<string> {
-    const transactionHash = await query(ethQuery, 'sendRawTransaction', [
-      transactionMeta.rawTx,
-    ]);
+    try {
+      const transactionHash = await query(ethQuery, 'sendRawTransaction', [
+        transactionMeta.rawTx,
+      ]);
 
-    if (skipSubmitHistory !== true) {
-      this.#updateSubmitHistory(transactionMeta, transactionHash);
+      if (skipSubmitHistory !== true) {
+        this.#updateSubmitHistory(transactionMeta, transactionHash);
+      }
+
+      return transactionHash;
+    } catch (error: unknown) {
+      const errorObject = error as
+        | {
+            data?: { message?: string };
+            message?: string;
+          }
+        | undefined;
+
+      const errorMessage =
+        errorObject?.data?.message ?? errorObject?.message ?? String(error);
+
+      throw new Error(errorMessage);
     }
-
-    return transactionHash;
   }
 
   /**
@@ -4664,6 +4722,11 @@ export class TransactionController extends BaseController<
     this.messenger.registerActionHandler(
       `${controllerName}:estimateGas`,
       this.estimateGas.bind(this),
+    );
+
+    this.messenger.registerActionHandler(
+      `${controllerName}:estimateGasBatch`,
+      this.estimateGasBatch.bind(this),
     );
 
     this.messenger.registerActionHandler(