+++ packages/transaction-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add intent based transaction support ([#6963](https://github.com/MetaMask/core/pull/6963))
+
 ## [62.6.0]
 
 ### Added
@@ -17,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add optional `isTimeoutEnabled` callback to disable for specific transactions.
   - Ignores transactions with future nonce.
   - Threshold determined by feature flag.
-- Add intent based transaction support ([#6963](https://github.com/MetaMask/core/pull/6963))
 
 ### Changed